warehouse controller

+++ src/main/java/com/perezmejia/systemvd/controller/CategoryController.java
@@ -58,7 +58,7 @@ public class CategoryController {
     }
 
     @RequestMapping(value = EDIT_PATH, method = RequestMethod.GET)
-    public String edit(Model model, @RequestParam(value = "id", required = true) int id) {
+    public String edit(Model model, @RequestParam(value = "id", required = true) Long id) {
         CategoryView categoryView = categoryService.findById(id);
 
         model.addAttribute("breadcrumb", messageSource.getMessage("breadcrumb.edit", null, Locale.getDefault()));
