+++ src/main/java/com/perezmejia/systemvd/controller/ClientController.java
@@ -66,7 +66,7 @@ public class ClientController {
     }
 
     @RequestMapping(value = EDIT_PATH, method = RequestMethod.GET)
-    public String edit(Model model, @RequestParam(value = "id", required = true) int id) {
+    public String edit(Model model, @RequestParam(value = "id", required = true) Long id) {
         ClientView clientView = clientService.findById(id);
 
         model.addAttribute("breadcrumb", messageSource.getMessage("breadcrumb.edit", null, Locale.getDefault()));
