+++ src/main/java/com/perezmejia/systemvd/service/inventory/impl/CategoryServiceImpl.java
@@ -28,7 +28,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public CategoryView findById(Integer id) {
+    public CategoryView findById(Long id) {
         return new CategoryView(categoryRepository.findOne(id));
     }
